% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleganza.slice.R
\name{colleganza.slice}
\alias{colleganza.slice}
\title{Create social networks with slices from the colleganza dataset}
\usage{
colleganza.slice(from = 1073, to = 1342)
}
\arguments{
\item{from}{Year of the first contract, 1073 by default}

\item{to}{Year of the last contract, 1342 by default, which is the last in the dataset}
}
\value{
A graph with with families as nodes, co-occurrence in a contract as edges, but only for the contracts between those dates
}
\description{
Create social networks with slices from the colleganza Dataset
}

\references{
\insertAllCited{}
}

\examples{
library(serrata.families)
library(igraph)
# There is a gap between 1261 and 1310
contracts.pre.1261 <- colleganza.slice(to=1261)
plot(contracts.pre.1261)

# The serrata took place between 1295 and 1330, but the dataset goes until 1342
contracts.post.1310 <- colleganza.slice(from=1310)
plot(contracts.post.1310)
}
